Desktop/genai-finance-chatbot/requirements.txt
python-dotenv==1.1.0
llama-index
llama-index-llms-huggingface-api
sentence-transformers
chromadb
python-dotenv
numpy
requests
huggingface_hub
